+++ src/components/Payments/Payments.tsx
@@ -1,5 +1,5 @@
 import { INPUT_TYPE } from "../../constants/constants";
-import InputForm from "../InputForm/InputForm";
+import InputSection from "../InputForm/InputSection";
 import Preview from "../Preview/Preview";
 import { PaymentsCSS } from "./Payments.styled";
 
@@ -7,9 +7,9 @@ function Payments() {
   return (
     <PaymentsCSS>
       <Preview />
-      <InputForm type={INPUT_TYPE.cardNumbers} />
-      <InputForm type={INPUT_TYPE.expirationPeriod} />
-      <InputForm type={INPUT_TYPE.cvcNumber} />
+      <InputSection type={INPUT_TYPE.cardNumbers} />
+      <InputSection type={INPUT_TYPE.expirationPeriod} />
+      <InputSection type={INPUT_TYPE.cvcNumber} />
     </PaymentsCSS>
   );
 }
